requirements.txt
asgiref==3.8.1
Django==5.2.1
django-jalali-date==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
jalali_core==1.0.0
jdatetime==5.2.0
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
